# api/openapi/contract.base.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "CPRa SDK draft management contract",
    "version": "0.1.0-rc.1",
    "description": "Draft contract. V2 server qualification is pending. Existing v1 is separate."
  },
  "paths": {
...
    "/api/v2/operations/{id}/reselection": {
      "post": {
        "operationId": "CreateCollectionReselection",
        "summary": "Create a private original-file reselection attempt",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. Verification and resume are separate explicit steps; neither activates configuration.",
        "x-cpra-availability": "implemented",
        "responses": {
          "201": {
            "description": "Create a private original-file reselection attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/CollectionReselectionAttempt"
                }
              }
            }
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./models.json#/components/schemas/CollectionReselectionCreateRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/operations/{id}/reselection/{attempt}": {
      "get": {
        "operationId": "GetCollectionReselection",
        "summary": "Read bounded attempt progress without renewing expiry",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. Verification and resume are separate explicit steps; neither activates configuration.",
        "x-cpra-availability": "implemented",
        "responses": {
          "200": {
            "description": "Read bounded attempt progress without renewing expiry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/CollectionReselectionAttempt"
                }
              }
            }
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "maxLength": 36,
              "description": "Opaque attempt identity. This is not authorization."
            }
          }
        ]
      },
      "delete": {
        "operationId": "DiscardCollectionReselection",
        "summary": "Discard only this attempt and its temporary input",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. DELETE removes only disposable attempt state, never the original operation or any already committed prefix.",
        "x-cpra-availability": "implemented",
        "responses": {
          "204": {
            "description": "Discard only this attempt and its temporary input"
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "maxLength": 36,
              "description": "Opaque attempt identity. This is not authorization."
            }
          }
        ]
      }
    },
    "/api/v2/operations/{id}/reselection/{attempt}/sources/{source}": {
      "put": {
        "operationId": "UploadCollectionReselectionSource",
        "summary": "Stage the next encrypted source part",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. Verification and resume are separate explicit steps; neither activates configuration.",
        "x-cpra-availability": "implemented",
        "responses": {
          "200": {
            "description": "Stage the next encrypted source part",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/CollectionReselectionAttempt"
                }
              }
            }
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "maxLength": 36,
              "description": "Opaque attempt identity. This is not authorization."
            }
          },
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "maximum": 67108864
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "At most 1 MiB of exact raw file bytes. An empty body is allowed only when end=true. No multipart metadata, file names, URL fetching, archive extraction, JSON or base64 encoding. Total source input is bounded to 64 MiB. Only an identical retry of the most recently accepted part can reconcile a lost reply.",
          "x-cpra-max-body-bytes": 1048576,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 1048576
              }
            }
          }
        }
      }
    },
    "/api/v2/operations/{id}/reselection/{attempt}/verify": {
      "post": {
        "operationId": "VerifyCollectionReselection",
        "summary": "Admit asynchronous original-input verification",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. Verification and resume are separate explicit steps; neither activates configuration.",
        "x-cpra-availability": "implemented",
        "responses": {
          "202": {
            "description": "Admit asynchronous original-input verification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/CollectionReselectionAttempt"
                }
              }
            }
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "maxLength": 36,
              "description": "Opaque attempt identity. This is not authorization."
            }
          }
        ]
      }
    },
    "/api/v2/operations/{id}/reselection/{attempt}/resume": {
      "post": {
        "operationId": "ResumeCollectionReselection",
        "summary": "Admit asynchronous verified remainder upload",
        "description": "Available when the server configures collection reselection; discover support before use. The authenticated caller must remain the original operation owner. No request renews the original operation deadline. Verification and resume are separate explicit steps; neither activates configuration.",
        "x-cpra-availability": "implemented",
        "responses": {
          "202": {
            "description": "Admit asynchronous verified remainder upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/CollectionReselectionAttempt"
                }
              }
            }
          },
          "default": {
            "description": "Problem details. No private source data or commitments are returned.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "maxLength": 36,
              "description": "Opaque attempt identity. This is not authorization."
            }
          }
        ]
      }
    }
  }
}